+++ lib/rails_wayback/configuration.rb
@@ -10,14 +10,16 @@ module RailsWayback
     DEFAULT_VIEW_PATHS = ["app/views", "app/components"].freeze
     DEFAULT_ASSET_PATHS = ["app/assets", "public"].freeze
     DEFAULT_MAX_COMMITS = 50
+    DEFAULT_ALLOWED_ENVIRONMENTS = %w[development test].freeze
 
-    attr_accessor :view_paths, :asset_paths, :max_commits,
+    attr_accessor :view_paths, :asset_paths, :max_commits, :allowed_environments,
                   :app_root, :cache_root, :toggle_file
 
     def initialize
       @view_paths = DEFAULT_VIEW_PATHS.dup
       @asset_paths = DEFAULT_ASSET_PATHS.dup
       @max_commits = DEFAULT_MAX_COMMITS
+      @allowed_environments = DEFAULT_ALLOWED_ENVIRONMENTS.dup
       @app_root = nil
       @cache_root = nil
       @toggle_file = nil